[feat] Add token contexts
