+++ package.json
@@ -15,6 +15,7 @@
     "check-all": "yarn types:check && yarn format:check && yarn lint:check && yarn unused:check"
   },
   "dependencies": {
+    "@tanstack/react-query": "5.59.3",
     "react": "18.3.1",
     "react-dom": "18.3.1",
     "react-router-dom": "6.26.2"
