+++ src/context/TokenManageContext.ts
@@ -1,5 +1,6 @@
 import { createContext } from 'react';
 export type TokenManageContext = {
+  getToken: () => string | null;
   saveToken: (token: string) => void;
   clearToken: () => void;
 };
