+++ src/pages/landing/Landing.tsx
@@ -1,57 +1,19 @@
-import { useState } from 'react';
-
-import { ServiceContext } from '../../context/ServiceContext';
-import { useGaurdContext } from '../../hooks/useGuardContext';
+import { useNavigate } from 'react-router-dom';
 
 export const Landing = () => {
-  // 로그인 목업
-  const [id, setId] = useState('');
-  const [password, setPassword] = useState('');
+  const navigate = useNavigate();
 
-  const { authService } = useGaurdContext(ServiceContext);
-
-  const onClickButton = async () => {
-    const response = await authService.signIn({ id, password });
-    if (response.type === 'success') {
-      localStorage.setItem('token', response.data.token);
-      alert(`로그인 성공! 토큰은 ${response.data.token}입니다.`);
-    } else alert(response.message);
+  const onClickSignIn = () => {
+    navigate('/signIn');
   };
 
+  const onClickSignUp = () => {
+    navigate('signUp');
+  };
   return (
     <div>
-      <h1>Login</h1>
-      <div>
-        <label htmlFor="id">ID:</label>
-        <input
-          type="text"
-          id="id"
-          value={id}
-          onChange={(e) => {
-            setId(e.target.value);
-          }}
-        />
-      </div>
-      <div>
-        <label htmlFor="password">Password:</label>
-        <input
-          type="password"
-          id="password"
-          value={password}
-          onChange={(e) => {
-            setPassword(e.target.value);
-          }}
-        />
-      </div>
-      <button
-        onClick={() => {
-          onClickButton().catch(() => {
-            console.error('error');
-          });
-        }}
-      >
-        Submit
-      </button>
+      <button onClick={onClickSignIn}>로그인</button>
+      <button onClick={onClickSignUp}>회원가입</button>
     </div>
   );
 };
